requirements.txt
pydantic>=2.11.7,<2.12.0
PyGithub>=2.6.1,<2.7.0
python-slugify>=8.0.4,<8.1.0
PyYAML>=6.0.2,<6.1.0
PyJWT>=2.10.1,<2.11.0